% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{addRMessageHandler}
\alias{addRMessageHandler}
\title{Supply the name of a function to call, identified by its key}
\usage{
addRMessageHandler(key, functionName)
}
\arguments{
\item{key}{A character string}

\item{functionName}{A character string}
}
\description{
Supply the name of a function to call, identified by its key
}
