% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{data}
\name{BrowserViz-class}
\alias{BrowserViz-class}
\alias{BrowserViz.state}
\alias{BrowserViz}
\title{Constructor for BrowserViz}
\format{An object of class \code{environment} of length 1.}
\usage{
BrowserViz.state

BrowserViz(portRange = 10000:10100, title = "BrowserViz", browserFile,
  quiet = TRUE, httpQueryProcessingFunction = NULL)
}
\arguments{
\item{portRange}{The constructor looks for a free websocket port in this range.  15000:15100 by default}

\item{title}{Used for the web browser window, "igvR" by default}

\item{browserFile}{The full path to the bundled html, js and libraries, and css which constitute the browser app}

\item{quiet}{A logical variable controlling verbosity during execution}

\item{httpQueryProcessingFunction}{a function, default NULL, provides subclasses with the
opportunity to execute code on the http server created here.}
}
\value{
An object of the BrowserViZ class
}
\description{
asks your browser to display browserFile, which is presented
by the minimal http server offered by httpuv, after which websocket messages are
exchanged.  the default browserFile, viz.html, does not do much, but is copied and
extended by BrowserViz subclassing applcations
the optional sixth argument, httpQueryProcessingFunction,   one case of this is
the BrowserTable class, which uses http in an ajax-ish way to pass pages of a possibly
very large data.frame to the browser for incremental display.
}
\keyword{datasets}
