% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{dispatchMessage}
\alias{dispatchMessage}
\title{Route the message coming in from the browser to the appropriate R function.}
\usage{
dispatchMessage(ws, msg, quiet)
}
\arguments{
\item{ws}{a websocket connectin}

\item{msg}{the JSON-encoded message from the browser}

\item{quiet}{logical TRUE or FALSE}
}
\description{
Route the message coming in from the browser to the appropriate R function.
}
