% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{fromJSON}
\alias{fromJSON}
\title{Transform JSON string into a native R object}
\usage{
fromJSON(...)
}
\arguments{
\item{...}{Extra arguments passed to this function}
}
\value{
a native R data structure
}
\description{
Transform JSON string into a native R object
}
\examples{

 fromJSON(toJSON(data.frame(a=8:10, b=LETTERS[8:10], stringsAsFactors=FALSE)))

}
