% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{getBrowserInfo,BrowserViz-method}
\alias{getBrowserInfo,BrowserViz-method}
\alias{getBrowserInfo}
\title{Retrieve basic attributes of the attached web browser.}
\usage{
\S4method{getBrowserInfo}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\description{
Retrieve basic attributes of the attached web browser.
}
