% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{getBrowserResponse,BrowserViz-method}
\alias{getBrowserResponse,BrowserViz-method}
\alias{getBrowserResponse}
\title{Retrieve the response sent by the browser}
\usage{
\S4method{getBrowserResponse}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\description{
Retrieve the response sent by the browser
}
\examples{
library(BrowserViz)
browserVizBrowserFile <- system.file(package="BrowserViz", "browserCode", "dist", "bvDemoApp.html")
if(BrowserViz::webBrowserAvailableForTesting()){
  bvApp <- BrowserViz(browserFile=browserVizBrowserFile, quiet=TRUE)
  data <- list(lowercase=letters, uppercase=LETTERS)
  json.returned <- roundTripTest(bvApp, data)
  data.returned <- fromJSON(json.returned)
  stopifnot(identical(data, data.returned))
  closeWebSocket(bvApp)
  }

}
