% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{getBrowserWindowSize,BrowserViz-method}
\alias{getBrowserWindowSize,BrowserViz-method}
\alias{getBrowserWindowSize}
\title{Supply the name of a function to call, identified by its key}
\usage{
\S4method{getBrowserWindowSize}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\description{
Supply the name of a function to call, identified by its key
}
