% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{handleResponse}
\alias{handleResponse}
\title{handleResponse}
\usage{
handleResponse(ws, msg)
}
\arguments{
\item{ws}{websocket connection}

\item{msg}{the JSON-encoded character string returned by the browser}
}
\description{
handleResponse
}
