% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{ready,BrowserViz-method}
\alias{ready,BrowserViz-method}
\alias{ready}
\title{Is the websocket connection to the browser ready for use?}
\usage{
\S4method{ready}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\description{
Is the websocket connection to the browser ready for use?
}
