% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{roundTripTest,BrowserViz-method}
\alias{roundTripTest,BrowserViz-method}
\alias{roundTripTest}
\title{Send data to the browser, ensure that it is returned accurately.}
\usage{
\S4method{roundTripTest}{BrowserViz}(obj, ...)
}
\arguments{
\item{obj}{An object of class BrowserViz}

\item{...}{other arguments}
}
\description{
Send data to the browser, ensure that it is returned accurately.
}
\examples{
library(BrowserViz)
browserVizBrowserFile <- system.file(package="BrowserViz", "browserCode", "dist", "bvDemoApp.html")
if(BrowserViz::webBrowserAvailableForTesting()){
  bvApp <- BrowserViz(browserFile=browserVizBrowserFile, quiet=TRUE)
  data <- list(lowercase=letters, uppercase=LETTERS)
  json.returned <- roundTripTest(bvApp, data)
  data.returned <- fromJSON(json.returned)
  stopifnot(identical(data, data.returned))
  closeWebSocket(bvApp)
  }

}
