% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{send,BrowserViz-method}
\alias{send,BrowserViz-method}
\alias{send}
\title{Send the specified message to the browser}
\usage{
\S4method{send}{BrowserViz}(obj, msg)
}
\arguments{
\item{obj}{An object of class BrowserViz}

\item{msg}{A list with four fields: {cmd: "someCommand", status: "request",
callback: "someFunction", payload: "someData"}}
}
\description{
Send the specified message to the browser
}
