% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{setBrowserWindowTitle,BrowserViz-method}
\alias{setBrowserWindowTitle,BrowserViz-method}
\alias{setBrowserWindowTitle}
\title{Supply the name of a function to call, identified by its key}
\usage{
\S4method{setBrowserWindowTitle}{BrowserViz}(obj, newTitle)
}
\arguments{
\item{obj}{An object of class BrowserViz}

\item{newTitle}{A character string}
}
\description{
Supply the name of a function to call, identified by its key
}
