% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{show,BrowserViz-method}
\alias{show,BrowserViz-method}
\alias{show}
\title{Display the core attributes of the BrowserViz object to stdout}
\usage{
\S4method{show}{BrowserViz}(object)
}
\arguments{
\item{object}{An object of class BrowserViz}
}
\description{
Display the core attributes of the BrowserViz object to stdout
}
