% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{toJSON}
\alias{toJSON}
\title{Transform an R data structure into JSON}
\usage{
toJSON(..., auto_unbox = TRUE)
}
\arguments{
\item{...}{Extra arguments passed to this function}

\item{auto_unbox}{Logical}
}
\value{
a character string with the JSON representation of the R object
}
\description{
The semantics of toJSON changed between RJSONIO and jsonlite: in the latter, scalars are
promoted to arrays of length 1.  rather than change our javascript code, and since such
promotion -- while sensible in the context of R -- strikes me as gratuitous, I follow
jeroen ooms suggestion, creating this wrapper
}
\examples{

 toJSON(data.frame(a=8:10, b=LETTERS[8:10], stringsAsFactors=FALSE))

}
