% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{webBrowserAvailableForTesting}
\alias{webBrowserAvailableForTesting}
\title{Is there a web browser available for testing?}
\usage{
webBrowserAvailableForTesting()
}
\value{
Logical TRUE or FALSE
}
\description{
This package's unit tests require a web browser to connect to.  our heuristic, though not bullet
 proof, is that one of three conditions must be met
Supply the name of a function to call, identified by its key
}
