% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{browserResponseReady,BrowserViz-method}
\alias{browserResponseReady,BrowserViz-method}
\alias{browserResponseReady}
\title{browserResponseReady}
\usage{
\S4method{browserResponseReady}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\description{
browserResponseReady
}
