% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{closeWebSocket,BrowserViz-method}
\alias{closeWebSocket,BrowserViz-method}
\alias{closeWebSocket}
\title{Close the websocket connection - between your R session and your web browser.}
\usage{
\S4method{closeWebSocket}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\description{
Close the websocket connection - between your R session and your web browser.
}
