% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\name{displayHTMLInDiv,BrowserViz-method}
\alias{displayHTMLInDiv,BrowserViz-method}
\alias{displayHTMLInDiv}
\title{Ask the browser to display html markup in the specified div}
\usage{
\S4method{displayHTMLInDiv}{BrowserViz}(obj, htmlText, div.id)
}
\arguments{
\item{obj}{An object of class BrowserViz}

\item{htmlText}{A character string with HTML markup}

\item{div.id}{A character string}
}
\description{
Ask the browser to display html markup in the specified div
}
