% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{getBrowserResponse,BrowserViz-method}
\alias{getBrowserResponse,BrowserViz-method}
\alias{getBrowserResponse}
\title{Retrieve the response sent by the browser}
\usage{
\S4method{getBrowserResponse}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\description{
Retrieve the response sent by the browser
}
