% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{port,BrowserViz-method}
\alias{port,BrowserViz-method}
\alias{port}
\title{Get the port number}
\usage{
\S4method{port}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\value{
the port number use in the websocket connection, a numeric value.
}
\description{
Get the port number
}
