% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{roundTripTest,BrowserViz-method}
\alias{roundTripTest,BrowserViz-method}
\alias{roundTripTest}
\title{Supply the name of a function to call, identified by its key}
\usage{
\S4method{roundTripTest}{BrowserViz}(obj, ...)
}
\arguments{
\item{obj}{An object of class BrowserViz}

\item{...}{other arguments}
}
\description{
Supply the name of a function to call, identified by its key
}
