% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserViz-class.R
\docType{methods}
\name{show,BrowserViz-method}
\alias{show,BrowserViz-method}
\alias{show}
\title{Display to sdtout the core attributes of the BrowserViz object}
\usage{
\S4method{show}{BrowserViz}(obj)
}
\arguments{
\item{obj}{An object of class BrowserViz}
}
\description{
Display to sdtout the core attributes of the BrowserViz object
}
